src/app/backend/requirements.txt
uuid
opencv-python
fastapi
numpy
pillow
keras
uvicorn